+++ tests/integration_tests/test_tracers.py
@@ -2,6 +2,7 @@ import itertools
 import json
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
+import pytest
 from web3 import Web3
 
 from .expected_constants import (
@@ -24,6 +25,7 @@ from .utils import (
     send_txs,
     sign_transaction,
     w3_wait_for_new_blocks,
+    wait_for_fn,
 )
 
 
@@ -189,6 +191,56 @@ def test_trace_tx_reverse_transfer(ethermint):
             [tx_hash, tracer],
         )
         print(tx_res)
+
+
+@pytest.mark.flaky(max_runs=10)
+def test_destruct(ethermint):
+    method = "debug_traceTransaction"
+    tracer = {"tracer": "callTracer"}
+    receiver = "0x0F0cb39319129BA867227e5Aae1abe9e7dd5f861"
+    acc = derive_new_account(11)  # ethm13c2n7geavjfsqcan290mq74kajjlxehyzhly4p
+    w3 = ethermint.w3
+    fund_acc(w3, acc, fund=3077735635376769427)
+    sender = acc.address
+    raw_transactions = []
+    contracts = []
+    total = 3
+    for _ in range(total):
+        contract, _ = deploy_contract(w3, CONTRACTS["SelfDestruct"], key=acc.key)
+        contracts.append(contract)
+
+    nonce = w3.eth.get_transaction_count(sender)
+
+    for i in range(total):
+        tx = (
+            contracts[i]
+            .functions.execute()
+            .build_transaction(
+                {
+                    "from": sender,
+                    "nonce": nonce,
+                    "gas": 167115,
+                    "gasPrice": 5050000000000,
+                    "value": 353434350000000000,
+                }
+            )
+        )
+        raw_transactions.append(sign_transaction(w3, tx, acc.key).rawTransaction)
+        nonce += 1
+    sended_hash_set = send_raw_transactions(w3, raw_transactions)
+
+    def wait_balance():
+        return w3.eth.get_balance(receiver) > 0
+
+    wait_for_fn("wait_balance", wait_balance)
+    for h in sended_hash_set:
+        tx_hash = h.hex()
+        res = w3.provider.make_request(
+            method,
+            [tx_hash, tracer],
+        )
+        print(tx_hash, res)
+        assert "insufficient funds" not in res, res
 
 
 def test_tracecall_insufficient_funds(ethermint, geth):